Get rid of the chained resource transformer
