--- controller/src/main/java/org/jboss/as/controller/transform/chained/ChainedResourceTransformerEntry.java
@@ -1,45 +1,0 @@
-/*
-* JBoss, Home of Professional Open Source.
-* Copyright 2012, Red Hat Middleware LLC, and individual contributors
-* as indicated by the @author tags. See the copyright.txt file in the
-* distribution for a full listing of individual contributors.
-*
-* This is free software; you can redistribute it and/or modify it
-* under the terms of the GNU Lesser General Public License as
-* published by the Free Software Foundation; either version 2.1 of
-* the License, or (at your option) any later version.
-*
-* This software is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-* Lesser General Public License for more details.
-*
-* You should have received a copy of the GNU Lesser General Public
-* License along with this software; if not, write to the Free
-* Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-* 02110-1301 USA, or see the FSF site: http://www.fsf.org.
-*/
-package org.jboss.as.controller.transform.chained;
-
-import org.jboss.as.controller.OperationFailedException;
-import org.jboss.as.controller.PathAddress;
-import org.jboss.as.controller.registry.Resource;
-import org.jboss.as.controller.transform.ResourceTransformationContext;
-import org.jboss.as.controller.transform.ResourceTransformer;
-import org.jboss.as.controller.transform.description.TransformationDescriptionBuilder;
-
-/**
- * Provides chained transformation of resources.
- *
- * @deprecated Use {@link TransformationDescriptionBuilder} instead
- * @author Kabir Khan
- */
-@Deprecated
-public interface ChainedResourceTransformerEntry {
-
-    /**
-     * Similar to {@link ResourceTransformer#transformResource(ResourceTransformationContext, PathAddress, Resource)} but for use with {@link ChainedOperationTransformer}
-     */
-    @SuppressWarnings("deprecation")
-    void transformResource(ChainedResourceTransformationContext context, PathAddress address, Resource resource) throws OperationFailedException;
-}
